refactor(node/state_delta): extract events submodule (increment 2)
Continues the state_delta god-file split. Moves the event-execution
cluster into `events.rs` — behavior-preserving, pure code movement
(mod.rs: +5 wiring lines, -292).

Moved (all internal, re-imported into mod.rs so call sites are unchanged):
- `parse_events_payload` — JSON → ExecutionEvent decoding
- `execute_event_handlers_parsed` — run app handlers for a payload
- `execute_cascaded_events` — handler execution for peer-cascaded deltas
- `emit_state_mutation_event_parsed` — push StateMutation to WebSocket
- `CascadeOutcome` — result struct (fields `pub(super)` for mod.rs reads)

The `parse_events_payload` unit tests stay in mod.rs's test module and
resolve through the re-import. No logic changes.

Validated: node lib tests (274) and sync_sim (314) pass.

Stacked on #2697 (crypto + verify). Remaining: buffering/replay and
delta-store setup.
